+++ src/lookup/mod.rs
@@ -1,7 +1,9 @@
 pub mod asn;
 pub mod geo;
+pub mod ix;
 pub mod rdns;
 
 pub use asn::*;
 pub use geo::*;
+pub use ix::*;
 pub use rdns::*;